requirements.txt
streamlit==1.32.0
yfinance==0.2.36
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
openai==1.12.0
python-dotenv==1.0.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
python-docx==1.1.0
openpyxl==3.1.2
ta==0.11.0